+++ posthog-react-native/CHANGELOG.md
@@ -1,6 +1,7 @@
 ## Next
 
-1. if `react-native-device-info` is available for the Web target, do not set `unknown` for all properties.
+1. Do not try to load the `expo-file-system` package on the Web target since it's not supported.
+2. if `react-native-device-info` is available for the Web target, do not set `unknown` for all properties.
 
 # 2.10.1 - 2024-01-15